Git-aware versioning
Versions for both Mozilla and OpenOffice.org products are based on the name of the latest annotated tag associated with an ancestor of the current commit. If the current commit is ahead of the tag, the number of changes ahead is appended to the version string. The date variables are also taken from the date of the latest annotated tag.